# artifacts/api-server/src/routes/zana.ts
import { Router, type IRouter } from "express";
import { and, desc, eq } from "drizzle-orm";
import { randomUUID } from "node:crypto";
import { db, activityTable, projectMembersTable, projectsTable, tasksTable, usersTable } from "@workspace/db";
import {
  CreateInviteBody,
  CreateInviteParams,
  CreateInviteResponse,
  CreateProjectBody,
  CreateProjectResponse,
  CreateTaskBody,
  CreateTaskParams,
  CreateTaskResponse,
  DeleteProjectParams,
  DeleteTaskParams,
  GetMeResponse,
  GetOverviewResponse,
  GetProjectParams,
  GetProjectResponse,
  ListProjectMembersParams,
  ListProjectMembersResponse,
  ListProjectsResponse,
  UpdateProjectBody,
  UpdateProjectParams,
  UpdateProjectResponse,
  UpdateTaskBody,
  UpdateTaskParams,
  UpdateTaskResponse,
} from "@workspace/api-zod";

const router: IRouter = Router();
const CURRENT_USER_ID = "user-jordan";
const DEFAULT_PROJECT_ID = "project-launch";
let seedPromise: Promise<void> | null = null;

const now = () => new Date();
const iso = (value: Date | string) => value instanceof Date ? value.toISOString() : value;
const initialsFor = (value: string) =>
  value.split(/[.\s_-]+/).filter(Boolean).slice(0, 2).map((part) => part[0]?.toUpperCase()).join("") || "NA";

async function ensureSeed(): Promise<void> {
  const [user] = await db.select().from(usersTable).where(eq(usersTable.id, CURRENT_USER_ID));
  if (user) return;

  if (seedPromise) {
    await seedPromise;
    return;
  }

  seedPromise = (async () => {
    const [existingUser] = await db.select().from(usersTable).where(eq(usersTable.id, CURRENT_USER_ID));
    if (existingUser) return;

    await db.insert(usersTable).values({
    id: CURRENT_USER_ID,
    name: "Jordan Lee",
    email: "jordan@zana.app",
    initials: "JL",
    });
    await db.insert(projectsTable).values([
    {
      id: DEFAULT_PROJECT_ID,
      name: "Product launch",
      description: "Everything needed to take the next release from idea to shipped.",
      color: "ink",
    },
    {
      id: "project-studio",
      name: "Studio refresh",
      description: "A sharper home for the studio and the work behind it.",
      color: "stone",
    },
    {
      id: "project-writing",
      name: "Writing practice",
      description: "A small, steady system for publishing more often.",
      color: "paper",
    },
    ]);
    await db.insert(projectMembersTable).values([
    { id: "member-jordan-launch", projectId: DEFAULT_PROJECT_ID, userId: CURRENT_USER_ID, email: "jordan@zana.app", name: "Jordan Lee", initials: "JL", role: "owner", status: "active" },
    { id: "member-maya-launch", projectId: DEFAULT_PROJECT_ID, userId: "user-maya", email: "maya@zana.app", name: "Maya Chen", initials: "MC", role: "member", status: "active" },
    { id: "member-omar-launch", projectId: DEFAULT_PROJECT_ID, userId: "user-omar", email: "omar@zana.app", name: "Omar Haddad", initials: "OH", role: "member", status: "active" },
    { id: "member-jordan-studio", projectId: "project-studio", userId: CURRENT_USER_ID, email: "jordan@zana.app", name: "Jordan Lee", initials: "JL", role: "owner", status: "active" },
    { id: "member-jordan-writing", projectId: "project-writing", userId: CURRENT_USER_ID, email: "jordan@zana.app", name: "Jordan Lee", initials: "JL", role: "owner", status: "active" },
    ]);
    await db.insert(tasksTable).values([
    { id: "task-research", projectId: DEFAULT_PROJECT_ID, title: "Research onboarding friction", description: "Talk to three people who joined last month and capture the common thread.", status: "not_done", assigneeId: "user-maya", position: 0 },
    { id: "task-brief", projectId: DEFAULT_PROJECT_ID, title: "Write the release brief", description: "Turn the scattered notes into one clear page the team can rally around.", status: "doing", assigneeId: CURRENT_USER_ID, position: 0 },
    { id: "task-demo", projectId: DEFAULT_PROJECT_ID, title: "Record product walkthrough", description: "A concise walkthrough for the first announcement.", status: "doing", assigneeId: "user-omar", position: 1 },
    { id: "task-pricing", projectId: DEFAULT_PROJECT_ID, title: "Confirm pricing page copy", description: "Final review with the words team.", status: "done", assigneeId: CURRENT_USER_ID, position: 0 },
    { id: "task-qa", projectId: DEFAULT_PROJECT_ID, title: "Run final QA pass", description: "Check the critical path on desktop and mobile.", status: "done", assigneeId: "user-maya", position: 1 },
    ]);
    await db.insert(activityTable).values([
    { id: "activity-1", text: "Maya moved Research onboarding friction to Not done", time: "12 min ago", kind: "task" },
    { id: "activity-2", text: "Jordan created a new project, Writing practice", time: "2 hours ago", kind: "project" },
    { id: "activity-3", text: "Omar joined Product launch", time: "Yesterday", kind: "member" },
    ]);
  })();

  try {
    await seedPromise;
  } finally {
    seedPromise = null;
  }
}

async function projectView(project: typeof projectsTable.$inferSelect) {
  const tasks = await db.select().from(tasksTable).where(eq(tasksTable.projectId, project.id));
  return {
    id: project.id,
    name: project.name,
    description: project.description,
    color: project.color,
    taskCount: tasks.length,
    updatedAt: iso(project.updatedAt),
  };
}

async function taskView(task: typeof tasksTable.$inferSelect) {
  const [member] = task.assigneeId
    ? await db.select().from(projectMembersTable).where(and(eq(projectMembersTable.projectId, task.projectId), eq(projectMembersTable.userId, task.assigneeId)))
    : [];
  return {
    id: task.id,
    projectId: task.projectId,
    title: task.title,
    description: task.description,
    status: task.status as "not_done" | "doing" | "done",
    assigneeId: task.assigneeId,
    assigneeName: member?.name ?? null,
    assigneeInitials: member?.initials ?? null,
    createdAt: iso(task.createdAt),
    updatedAt: iso(task.updatedAt),
  };
}

async function addActivity(text: string, kind: "task" | "project" | "member") {
  await db.insert(activityTable).values({ id: randomUUID(), text, time: "Just now", kind });
}

router.get("/me", async (_req, res): Promise<void> => {
  await ensureSeed();
  const [user] = await db.select().from(usersTable).where(eq(usersTable.id, CURRENT_USER_ID));
  res.json(GetMeResponse.parse(user));
});

router.get("/overview", async (_req, res): Promise<void> => {
  await ensureSeed();
  const projects = await db.select().from(projectsTable);
  const tasks = await db.select().from(tasksTable);
  const members = await db.select().from(projectMembersTable);
  const activities = await db.select().from(activityTable).orderBy(desc(activityTable.createdAt)).limit(6);
  res.json(GetOverviewResponse.parse({
    projectCount: projects.length,
    openTaskCount: tasks.filter((task) => task.status !== "done").length,
    completedTaskCount: tasks.filter((task) => task.status === "done").length,
    memberCount: new Set(members.map((member) => member.userId ?? member.email)).size,
    activities: activities.map((activity) => ({ id: activity.id, text: activity.text, time: activity.time, kind: activity.kind as "task" | "project" | "member" })),
  }));
});

router.get("/projects", async (_req, res): Promise<void> => {
  await ensureSeed();
  const projects = await db.select().from(projectsTable).orderBy(desc(projectsTable.updatedAt));
  res.json(ListProjectsResponse.parse(await Promise.all(projects.map(projectView))));
});

router.post("/projects", async (req, res): Promise<void> => {
  await ensureSeed();
  const parsed = CreateProjectBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const timestamp = now();
  const project = {
    id: randomUUID(),
    name: parsed.data.name,
    description: parsed.data.description ?? "",
    color: parsed.data.color ?? "ink",
    createdAt: timestamp,
    updatedAt: timestamp,
  };
  await db.insert(projectsTable).values(project);
  await db.insert(projectMembersTable).values({
    id: randomUUID(),
    projectId: project.id,
    userId: CURRENT_USER_ID,
    email: "jordan@zana.app",
    name: "Jordan Lee",
    initials: "JL",
    role: "owner",
    status: "active",
  });
  await addActivity(`Jordan created a new project, ${project.name}`, "project");
  res.status(201).json(CreateProjectResponse.parse(await projectView(project)));
});

router.get("/projects/:projectId", async (req, res): Promise<void> => {
  await ensureSeed();
  const parsed = GetProjectParams.safeParse(req.params);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [project] = await db.select().from(projectsTable).where(eq(projectsTable.id, parsed.data.projectId));
  if (!project) {
    res.status(404).json({ error: "Project not found" });
    return;
  }
  const tasks = await db.select().from(tasksTable).where(eq(tasksTable.projectId, project.id)).orderBy(tasksTable.position);
  const members = await db.select().from(projectMembersTable).where(eq(projectMembersTable.projectId, project.id));
  res.json(GetProjectResponse.parse({
    project: await projectView(project),
    tasks: await Promise.all(tasks.map(taskView)),
    members: members.map((member) => ({ id: member.id, name: member.name, email: member.email, initials: member.initials, role: member.role as "owner" | "member", status: member.status as "active" | "invited" })),
  }));
});

router.patch("/projects/:projectId", async (req, res): Promise<void> => {
  await ensureSeed();
  const params = UpdateProjectParams.safeParse(req.params);
  const parsed = UpdateProjectBody.safeParse(req.body);
  if (!params.success || !parsed.success) {
    res.status(400).json({ error: "Invalid project update" });
    return;
  }
  const [project] = await db.update(projectsTable).set({ ...parsed.data, updatedAt: now() }).where(eq(projectsTable.id, params.data.projectId)).returning();
  if (!project) {
    res.status(404).json({ error: "Project not found" });
    return;
  }
  res.json(UpdateProjectResponse.parse(await projectView(project)));
});

router.delete("/projects/:projectId", async (req, res): Promise<void> => {
  await ensureSeed();
  const params = DeleteProjectParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  await db.delete(tasksTable).where(eq(tasksTable.projectId, params.data.projectId));
  await db.delete(projectMembersTable).where(eq(projectMembersTable.projectId, params.data.projectId));
  const deleted = await db.delete(projectsTable).where(eq(projectsTable.id, params.data.projectId)).returning();
  if (!deleted[0]) {
    res.status(404).json({ error: "Project not found" });
    return;
  }
  res.sendStatus(204);
});

router.get("/projects/:projectId/members", async (req, res): Promise<void> => {
  await ensureSeed();
  const params = ListProjectMembersParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const members = await db.select().from(projectMembersTable).where(eq(projectMembersTable.projectId, params.data.projectId));
  res.json(ListProjectMembersResponse.parse(members.map((member) => ({ id: member.id, name: member.name, email: member.email, initials: member.initials, role: member.role as "owner" | "member", status: member.status as "active" | "invited" }))));
});

router.post("/projects/:projectId/tasks", async (req, res): Promise<void> => {
  await ensureSeed();
  const params = CreateTaskParams.safeParse(req.params);
  const parsed = CreateTaskBody.safeParse(req.body);
  if (!params.success || !parsed.success) {
    res.status(400).json({ error: "Invalid task" });
    return;
  }
  const timestamp = now();
  const task = {
    id: randomUUID(),
    projectId: params.data.projectId,
    title: parsed.data.title,
    description: parsed.data.description ?? "",
    status: parsed.data.status,
    assigneeId: parsed.data.assigneeId ?? null,
    createdAt: timestamp,
    updatedAt: timestamp,
    position: 0,
  };
  await db.insert(tasksTable).values(task);
  await db.update(projectsTable).set({ updatedAt: timestamp }).where(eq(projectsTable.id, task.projectId));
  await addActivity(`Jordan created ${task.title}`, "task");
  res.status(201).json(CreateTaskResponse.parse(await taskView(task)));
});

router.patch("/projects/:projectId/tasks/:taskId", async (req, res): Promise<void> => {
  await ensureSeed();
  const params = UpdateTaskParams.safeParse(req.params);
  const parsed = UpdateTaskBody.safeParse(req.body);
  if (!params.success || !parsed.success) {
    res.status(400).json({ error: "Invalid task update" });
    return;
  }
  const [existing] = await db.select().from(tasksTable).where(and(eq(tasksTable.id, params.data.taskId), eq(tasksTable.projectId, params.data.projectId)));
  if (!existing) {
    res.status(404).json({ error: "Task not found" });
    return;
  }
  const [task] = await db.update(tasksTable).set({ ...parsed.data, updatedAt: now() }).where(eq(tasksTable.id, existing.id)).returning();
  await db.update(projectsTable).set({ updatedAt: now() }).where(eq(projectsTable.id, existing.projectId));
  res.json(UpdateTaskResponse.parse(await taskView(task)));
});

router.delete("/projects/:projectId/tasks/:taskId", async (req, res): Promise<void> => {
  await ensureSeed();
  const params = DeleteTaskParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const deleted = await db.delete(tasksTable).where(and(eq(tasksTable.id, params.data.taskId), eq(tasksTable.projectId, params.data.projectId))).returning();
  if (!deleted[0]) {
    res.status(404).json({ error: "Task not found" });
    return;
  }
  res.sendStatus(204);
});

router.post("/projects/:projectId/invites", async (req, res): Promise<void> => {
  await ensureSeed();
  const params = CreateInviteParams.safeParse(req.params);
  const parsed = CreateInviteBody.safeParse(req.body);
  if (!params.success || !parsed.success) {
    res.status(400).json({ error: "A valid email is required" });
    return;
  }
  const email = parsed.data.email.trim().toLowerCase();
  const [existing] = await db.select().from(projectMembersTable).where(and(eq(projectMembersTable.projectId, params.data.projectId), eq(projectMembersTable.email, email)));
  if (existing) {
    res.status(409).json({ error: "This person is already part of the project" });
    return;
  }
  const name = email.split("@")[0].replace(/[._-]+/g, " ").replace(/\b\w/g, (letter) => letter.toUpperCase());
  const member = { id: randomUUID(), projectId: params.data.projectId, userId: null, email, name, initials: initialsFor(name), role: "member", status: "invited" };
  await db.insert(projectMembersTable).values(member);
  await addActivity(`Jordan invited ${email}`, "member");
  res.status(201).json(CreateInviteResponse.parse(member));
});

export default router;