# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Zana project management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  initials: string;
}

export interface Project {
  id: string;
  name: string;
  description: string;
  color: string;
  taskCount: number;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  color?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  color?: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  not_done: 'not_done',
  doing: 'doing',
  done: 'done',
} as const;

export interface Task {
  id: string;
  projectId: string;
  title: string;
  description: string;
  status: TaskStatus;
  /** @nullable */
  assigneeId: string | null;
  /** @nullable */
  assigneeName: string | null;
  /** @nullable */
  assigneeInitials: string | null;
  createdAt: string;
  updatedAt: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  not_done: 'not_done',
  doing: 'doing',
  done: 'done',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status: TaskInputStatus;
  /** @nullable */
  assigneeId?: string | null;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  not_done: 'not_done',
  doing: 'doing',
  done: 'done',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  /** @nullable */
  assigneeId?: string | null;
}

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  owner: 'owner',
  member: 'member',
} as const;

export type MemberStatus = typeof MemberStatus[keyof typeof MemberStatus];


export const MemberStatus = {
  active: 'active',
  invited: 'invited',
} as const;

export interface Member {
  id: string;
  name: string;
  email: string;
  initials: string;
  role: MemberRole;
  status: MemberStatus;
}

export interface ProjectBoard {
  project: Project;
  tasks: Task[];
  members: Member[];
}

export interface InviteInput {
  email: string;
}

export type ActivityKind = typeof ActivityKind[keyof typeof ActivityKind];


export const ActivityKind = {
  task: 'task',
  project: 'project',
  member: 'member',
} as const;

export interface Activity {
  id: string;
  text: string;
  time: string;
  kind: ActivityKind;
}

export interface Overview {
  projectCount: number;
  openTaskCount: number;
  completedTaskCount: number;
  memberCount: number;
  activities: Activity[];
}