# lib/api-zod/src/generated/types/overview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Zana project management API
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';

export interface Overview {
  projectCount: number;
  openTaskCount: number;
  completedTaskCount: number;
  memberCount: number;
  activities: Activity[];
}