# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Zana project management API
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: string;
  name: string;
  description: string;
  color: string;
  taskCount: number;
  updatedAt: string;
}