# lib/api-zod/src/generated/types/projectBoard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Zana project management API
 * OpenAPI spec version: 0.1.0
 */
import type { Member } from './member';
import type { Project } from './project';
import type { Task } from './task';

export interface ProjectBoard {
  project: Project;
  tasks: Task[];
  members: Member[];
}