# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Zana project management API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: string;
  projectId: string;
  title: string;
  description: string;
  status: TaskStatus;
  /** @nullable */
  assigneeId: string | null;
  /** @nullable */
  assigneeName: string | null;
  /** @nullable */
  assigneeInitials: string | null;
  createdAt: string;
  updatedAt: string;
}