# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Zana project management API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputStatus } from './taskInputStatus';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status: TaskInputStatus;
  /** @nullable */
  assigneeId?: string | null;
}